refactor: split anthropic stream adapter
